updateRestaurantForm

+++ backend/models/restaurant.py
@@ -40,7 +40,7 @@ class DBRestaurant(Base):
     longitude = Column(Float, nullable=False)
     street_name = Column(String, nullable=False)
     city = Column(String, nullable=False)
-    star_rating = Column(Float, nullable=True, default=None)
+    star_rating = Column(Integer, nullable=True, default=None)
     type = Column(sqlEnum(RestaurantType), nullable=True, default=RestaurantType.OTHER)
     radius_of_delivery_km = Column(Float, nullable=False, default=3)
     is_archived = Column(Boolean, nullable=True, default=False)
